Include URI for device identification (#18)
* Include URI for device identification

When receiving a call parse the SIP URI from the Contact or From header
to use for identifying a device. Using the URI will make it possible to
distinguish between multiple endpoints calling through a centralized
server (such as Asterisk) which would otherwise appear to come from the
same IP address. Keeping track of the Contact header URI will also make
it possible to send outbound calls to the device using the correct URI.

* Add action to run tests on PR

* Update github action

---------

Co-authored-by: Paulus Schoutsen <balloob@gmail.com>