Use header values for response host and port (#40)
Use the values from the Via header for determining the host and port to
send responses to. If the Via header can not be used fall back to the
Contact header. If both the Via header and Contact header can not be
used fall back to the peer host and port from the transport the message
was received on.